+++ server/source/Whiteboard.Server/Services/ICounter.cs
@@ -2,6 +2,6 @@ namespace Whiteboard.Server.Services
 {
     public interface ICounter
     {
-        int Next();
+        ulong Next();
     }
 }
